refactor(vestad): break the update module's two back edges
The update orchestrator called into serve.rs for the per-agent snapshot while
serve.rs called into it to start an update, and state.rs carried types owned by
update.rs while update.rs read state.rs. Both are now one-way: the per-agent
snapshot plus retention moves to maintenance.rs, which already owns which agents
a pass selects, and the phase vocabulary plus the operation slot move to a new
operation.rs that imports nothing.
