+++ vestad/src/state.rs
@@ -12,7 +12,7 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::{Mutex, RwLock};
 
 use crate::settings::{load_settings, Settings};
-use crate::{agent_status, docker, mobile_app, update};
+use crate::{agent_status, docker, mobile_app, operation};
 
 pub(crate) const PROXY_MAX_BODY_BYTES: usize = 10 * 1024 * 1024; // 10 MB
 
@@ -118,10 +118,10 @@ pub struct AppState {
     pub(crate) refresh_live: Mutex<HashMap<String, RefreshFamily>>,
     agent_locks: Mutex<HashMap<String, Arc<tokio::sync::RwLock<()>>>>,
     pub(crate) tunnel_url: Mutex<Option<String>>,
-    pub(crate) update_info: Mutex<Option<update::UpdateInfo>>,
-    /// The gateway's one operation slot (see update.rs): what an update is doing right now, and the
+    pub(crate) update_info: Mutex<Option<operation::UpdateInfo>>,
+    /// The gateway's one operation slot (see operation.rs): what an update is doing right now, and the
     /// lock that keeps a second update or a gateway restart from racing it.
-    pub(crate) update_operation: update::UpdateOperation,
+    pub(crate) update_operation: operation::UpdateOperation,
     pub(crate) http_client: reqwest::Client,
     pub(crate) settings: RwLock<Settings>,
     /// Revocable, per-service credentials the agent proxy accepts for a private service.
@@ -195,7 +195,7 @@ impl AppState {
                 agent_locks: Mutex::new(HashMap::new()),
                 tunnel_url: Mutex::new(tunnel_url),
                 update_info: Mutex::new(None),
-                update_operation: update::UpdateOperation::new(),
+                update_operation: operation::UpdateOperation::new(),
                 http_client,
                 settings: RwLock::new(settings),
                 service_keys: RwLock::new(crate::service_keys::load_store()),
@@ -232,6 +232,17 @@ impl AppState {
 }
 
 pub type SharedState = Arc<AppState>;
+
+/// Acquire the per-agent serialization lock as an owned write guard — the single
+/// owner of the `agent_lock(name).write_owned()` idiom every mutating agent
+/// handler repeats. Owned so it can be held across the rest of an `async move`
+/// (e.g. the spawned backup/restore pipelines) without borrowing `state`.
+pub(crate) async fn agent_write_guard(
+    state: &AppState,
+    name: &str,
+) -> tokio::sync::OwnedRwLockWriteGuard<()> {
+    state.agent_lock(name).await.write_owned().await
+}
 
 // --- Response helpers ---
